Refactor configuration management
Instead of using the shell environment, we use a dictionary.  This is
cleaner, allows us to explicitly define default variables in a single
place (instead of all across the codebase), allows us to override
values in tests in a cleaner way, and makes us flexible to different
ways of configuration later on.

Implements: AP-428