docker-compose: Use .env for most environment vars
This is much cleaner, and easier to maintain, than keeping the whole list
in the docker-compose.yml file.  The main reason we were doing that was
to ensure we didn't leak unnecessary secrets into the Docker environment,
but the secrets we *do* use there are of higher sensitivity anyway.  This
eases maintenance burden and does not significantly impact security.