Implement release/environment tracking in Langfuse
This adds two environment variables, ``EXTRA_VERSION`` and
``DEPLOYMENT_ID``, which control what is shown in Langfuse.

Developers are encouraged to run Willa locally using ``bin/dev``, which
will set the variables correctly and then start Compose.

The variables are not documented because they are not intended to be set
manually or in the .env file.

Implements: AP-457