ETL: Instrument embeddings with Langfuse
This adds an observation for the embedding side of the ETL pipeline.
The chunked documents are marked as the input, and the resultant
document IDs are marked as the output.  This allows for some level of
introspection via Langfuse, albeit not as ideal as we hoped.

Tracing in Langfuse is configurable via the ``ETL_TRACING`` environment
variable, which defaults to ``False``.

Additionally, this changes the ``load_pdfs`` routine to return the
document ID as a key with the list of Document objects.  This allows
for better logging of what document is being embedded and for tracing
in Langfuse per-document.

Finally, this changes ``run_pipeline`` to generate embeddings for each
document individually.  This improves traces as each trace corresponds
to a single document, and eases debugging when a document fails.

Implements: AP-456