SQL: Add 'migration' file
Allows for the database to be created without running Prisma.  The file
is based on the output of ``pg_dump --schema-only willa`` in a container
after cleaning up the syntax a bit and grouping related elements.

Adds a ``bin/dbinit`` script that can be run to create the initial
database structure.

Reorder the Dockerfile a bit to ensure files that are more likely to
change are in the more outer layers.

Ref: AP-412