web: Allow CASProvider config to be any source
By defining ``env`` in ``CASProvider``, the superclass implementation
of ``is_configured`` requires the variables to be set in the actual
environment.  Configuration can come from reading in ``/run/secrets``,
an ``.env`` file, or the environment.

This commit defines our own ``is_configured`` that checks ``CONFIG``
for the needed variables instead of relying on the superclass.

Additionally, add ``CHAINLIT_AUTH_SECRET`` to ``_NEEDS_ENVIRON``.

Ref: AP-412