web: Add a custom data layer retrieval method
In local development (Docker Compose / .env), we have a ``DATABASE_URL``
variable that is set.  We do not have that in staging/prod Swarm stacks.
This causes the Chainlit default Data Layer init method to return None,
because it doesn't know how to connect to the database.

This introduces a small wrapper using the documented ``data_layer``
callback, which will use ``DATABASE_URL`` if present, but fall back to
the Postgres environment variables we use in the Swarm.

Closes: AP-530